refactor: streamline module file creation and improve error handling for Python version check

+++ nl_rigging_tools_drag.py
@@ -20,22 +20,28 @@ def onMayaDroppedPythonFile(*args):
 
     """
     )
-    mod_name = "nl_rigging_tools"
-    cur_dir = os.path.dirname(__file__)
-    app_dir = os.environ["MAYA_APP_DIR"]
-    mod_dir = os.path.join(app_dir, "modules")
-    mod_path = os.path.join(mod_dir, f"{mod_name}.mod")
+    if sys.version_info[0] < 3:
+        mc.confirmDialog(t="Error", m="Sorry, this tool requires Python 3", b=["OK"])
+        return
 
-    # Create mod file
-    mod_content = f"+ {mod_name} any {cur_dir}\nscripts: .\n"
+    mod_name = "nl_rigging_tools"
+    mod_path = os.path.join(
+        os.environ["MAYA_APP_DIR"],
+        "modules",
+        f"{mod_name}.mod",
+    )
     try:
+        cur_dir = os.path.dirname(__file__)
+        content = f"+ {mod_name} any {cur_dir}\nscripts: .\n"
         with open(mod_path, "w") as f:
-            f.write(mod_content)
+            f.write(content)
+
         logging.info(f"Module file created : {mod_path}")
+
     except Exception as e:
         logging.error(f"Failed to create module file: {e}")
 
-    # Open UI
+    # Load the tool
     import nl_modules.nl_rigging_tools as nlRT
     from importlib import reload
